migrate useCreateEventType hook to web
