+++ apps/web/modules/event-types/hooks/useCreateEventType.ts
@@ -1,6 +1,4 @@
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useEffect, useState } from "react";
-import { useForm } from "react-hook-form";
+import { useState } from "react";
 import type { z } from "zod";
 
 import { createEventTypeInput } from "@calcom/features/eventtypes/lib/types";
@@ -8,33 +6,8 @@ import { useDebounce } from "@calcom/lib/hooks/useDebounce";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import { HttpError } from "@calcom/lib/http-error";
 import type { EventType } from "@calcom/prisma/client";
-import { SchedulingType } from "@calcom/prisma/enums";
-import { unlockedManagedEventTypeProps } from "@calcom/prisma/zod-utils";
 import { trpc } from "@calcom/trpc/react";
-
-export type CreateEventTypeFormValues = z.infer<typeof createEventTypeInput>;
-
-export const useCreateEventTypeForm = () => {
-  const form = useForm<CreateEventTypeFormValues>({
-    defaultValues: {
-      length: 15,
-    },
-    resolver: zodResolver(createEventTypeInput),
-  });
-
-  const schedulingTypeWatch = form.watch("schedulingType");
-  const isManagedEventType = schedulingTypeWatch === SchedulingType.MANAGED;
-
-  useEffect(() => {
-    if (isManagedEventType) {
-      form.setValue("metadata.managedEventConfig.unlockedFields", unlockedManagedEventTypeProps);
-    } else {
-      form.setValue("metadata", null);
-    }
-  }, [schedulingTypeWatch]);
-
-  return { form, isManagedEventType };
-};
+import { useCreateEventTypeForm } from "@calcom/atoms/hooks/event-types/private/useCreateEventTypeForm";
 
 export const useCreateEventType = (
   onSuccessMutation: (eventType: EventType) => void,
