+++ packages/platform/atoms/hooks/event-types/private/useCreateEventTypeForm.ts
@@ -1,0 +1,32 @@
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useEffect } from "react";
+import { useForm } from "react-hook-form";
+import type { z } from "zod";
+
+import { createEventTypeInput } from "@calcom/features/eventtypes/lib/types";
+import { SchedulingType } from "@calcom/prisma/enums";
+import { unlockedManagedEventTypeProps } from "@calcom/prisma/zod-utils";
+
+export type CreateEventTypeFormValues = z.infer<typeof createEventTypeInput>;
+
+export const useCreateEventTypeForm = () => {
+  const form = useForm<CreateEventTypeFormValues>({
+    defaultValues: {
+      length: 15,
+    },
+    resolver: zodResolver(createEventTypeInput),
+  });
+
+  const schedulingTypeWatch = form.watch("schedulingType");
+  const isManagedEventType = schedulingTypeWatch === SchedulingType.MANAGED;
+
+  useEffect(() => {
+    if (isManagedEventType) {
+      form.setValue("metadata.managedEventConfig.unlockedFields", unlockedManagedEventTypeProps);
+    } else {
+      form.setValue("metadata", null);
+    }
+  }, [schedulingTypeWatch]);
+
+  return { form, isManagedEventType };
+};
